# crates/aivyx-server/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aivyx Engine API",
    "version": "1.1.0",
    "description": "AI agent orchestration engine with multi-agent teams, A2A/MCP protocol support, and enterprise features.",
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" },
    { "url": "https://api.aivyx.ai", "description": "Production" }
  ],
...
  "paths": {
...
    "/auth/login": {
      "post": {
        "summary": "SSO login",
        "tags": ["Auth"],
        "security": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "provider": { "type": "string" },
                  "token": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "expires_at": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/auth/me": {
      "get": {
        "summary": "Current user info",
        "tags": ["Auth"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Current user information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_id": { "type": "string" },
                    "email": { "type": "string" },
                    "provider": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/audit": {
      "get": {
        "summary": "Recent audit entries",
        "tags": ["Audit"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Recent audit log entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/AuditEntry" }
                }
              }
            }
          }
        }
      }
    },
    "/audit/verify": {
      "post": {
        "summary": "Verify audit chain",
        "description": "Verify the integrity of the tamper-evident audit log chain.",
        "tags": ["Audit"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Verification result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": { "type": "boolean" },
                    "entries_checked": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/audit/search": {
      "get": {
        "summary": "Search audit",
        "tags": ["Audit"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Search query"
          },
          {
            "name": "event_type",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by event type"
          }
        ],
        "responses": {
          "200": {
            "description": "Matching audit entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/AuditEntry" }
                }
              }
            }
          }
        }
      }
    },
...
    "/channels": {
      "get": {
        "summary": "List channels",
        "tags": ["Channels"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of channels",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Channel" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create channel",
        "tags": ["Channels"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Channel" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Channel created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Channel" }
              }
            }
          }
        }
      }
    },
    "/channels/{name}": {
      "get": {
        "summary": "Get channel",
        "tags": ["Channels"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Channel name"
          }
        ],
        "responses": {
          "200": {
            "description": "Channel details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Channel" }
              }
            }
          },
          "404": {
            "description": "Channel not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update channel",
        "tags": ["Channels"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Channel name"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Channel" }
            }
          }
        },
        "responses": {
          "200": { "description": "Channel updated" }
        }
      },
      "delete": {
        "summary": "Delete channel",
        "tags": ["Channels"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Channel name"
          }
        ],
        "responses": {
          "204": { "description": "Channel deleted" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token or multi-tenant API key"
      }
    },
    "schemas": {
...
      "AuditEntry": {
        "type": "object",
        "properties": {
          "timestamp": { "type": "string", "format": "date-time", "description": "When the event occurred" },
          "event_type": { "type": "string", "description": "Type of audit event" },
          "details": { "type": "object", "description": "Event-specific details" }
        },
        "required": ["timestamp", "event_type"]
      },
      "Channel": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "description": "Unique channel name" },
          "adapter_type": { "type": "string", "description": "Adapter type (slack, discord, matrix, etc.)" }
        },
        "required": ["name", "adapter_type"]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "Error code" },
          "message": { "type": "string", "description": "Human-readable error message" }
        },
        "required": ["error", "message"]
      }
    }
  },
  "security": [
    { "bearerAuth": [] }
  ]
}